requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
awsgi2==1.0.2
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via flask
flask==3.1.2
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
libadvian==1.9.0
    # via awsgi2
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
textgrid==1.6.1
    # via -r requirements.in
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
